feat: moved main card contents into it's own component
